requirements.txt
flask
psycopg2-binary
waitress
python-dotenv
